V1/core/services/ai-worker/requirements.txt
# Core
openai>=1.12.0
python-dotenv==1.0.1

# LangChain & LangGraph (versões compatíveis com Python 3.12 e Pydantic 2.x)
langchain==0.3.7
langchain-community>=0.3.0
langchain-openai==0.2.3
langgraph==0.2.34
langsmith==0.1.129

# Message Queue
pika==1.3.2
aio-pika==9.3.1
boto3>=1.34.0

# Database & Vector Store
qdrant-client>=1.16.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1

# HTTP
httpx==0.26.0
requests==2.31.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytz==2024.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
